Added `core::fmt`-like API for formatted output.
Defined the `Formatter`, `DebugStruct`, `DebugTuple`, `DebugList`, `DebugSet` types,
which serve the same purpose as the types in `core::fmt`

Defined `const_display_fmt` and `const_debug_fmt` methods for `AsciiStr` and
PWrapper wrapping `&str`, integer types, and slices of those.

Removed the U8Array trait, the IsU8Array marker type, since it wasn't actually necessary.

Added the `0x` and `0b` prefixes for hexadecimal and binary formatting in alternate mode,
in StrWriter.

Added `StrWriter::write_ascii_repeated` method to repeat an ascii character an amount of times.

Added `StrWriter::{make_formatter, clear}` methods

Added test for debug formatting of struct with decimal, hexadecimal, and binary integers.

Added methods to FormattingFlags to set the individual formatting modes.

Added a margin to FormattingFlag, with methods to increment and decrement it.
Replaced the `display` and `debug` constructors with a NEW associated constant.
Replaced its IsAlternate field with a boolean.

Added a `try_` macro, for doing the equivalent of `?` in const fns.

Added `PWrapper::slice` function to construct a `PWrapper<&[_]>` from a reference to array.