Added nonexhaustive struct, variants, enums in `impl_debug` macro.
Defined hacks for automatically std tyoes in PWrapper, leaving non-std type unwrapped.

Defined the GetTypeKind trait,
and IsArrayKind, IsStdKind, IsNotStdKind, TypeKindMarker marker types.

The GetTypeKind trait allows a type to say whether the type is from the standard library or not,
wrapping type in PWrapper if it is.

Added automatic impl of GetTypeKind for every type in the `impl_debug` macro.

Added `coerce_to_fmt` macro to coerce a reference to a type that has a `const_*_fmt` method.