Removed `const_*_len` methods, moved length computation into `Formatter`.
Now `Formatter` can use two different backends:
    - ComputeStrLength: For computing the length of the formatted output
    - StrWriter: For writing the formatted output
then depending on which of those the `Formatter` is constructed with,
it'll serve the function of computing the lenght or writing the formatted output.

Now the formatter has all of the `write_*` methods from StrWriter,
and the length, capacity, make_formatter ones.

Removed the Result returned from the `Formatter::debug_*` methods,
and the `*Debug::{field, entry}` methods, returning an error only in `*Debug::finish`.

All of this means that users will only have to write one function,
and it'll be only a little less ergonomic than implementing the Debug/Display trait manually.