Defined ConstDebug derive macro, which implements the `const_debug_fmt` method
Added "derive" cargo feature, to enable the derive macro and the `syn` features necessary for it.

Added tests for the ConstDebug derive.

Added helper types for the derive macro.

Added macros for returning errors in the derive macro.

Made `AsciiStr` implement the formatting methods inside the `impl_fmt` macro.

Made the `try_` macro use an `if let` instead of a `match`,
since it's always used with `()` as the Ok variant.

Added the "only_new_tests" cargo feature,to speed up compilation of new tests by cfg-ing out older tests.