Rewrite to only use syn for ConstDebug derive.
Rewrote the proc macro crate to not use syn for parsing the formatting macros,
fixing the spans of compile-time errors in the process.

Defined a few types/traits for parsing token streams, special cased for the formatting macros.

Defined a syn-like `Error` type.

Removed the generic parameter of LinearResult, making it a `Result<(), crate::Error>`

Updated proc macro tests.

Changed most proc macro code to use `__cf_osRcTFl4A` as the name of the crate,
instead of passing a `pcrate` variable to everything.
(the trailing random characters are to ensure that no accidental name colisions happen)

Fixed test for the `try_` macro, it was relying on the "constant_time_as_str" feature.

Added test ensuring that raw string literal formatting strings work,
as well as unicode escapes in regular string literals.