Made concatcp and formatcp use const generics when available.
Made the length computation always happen in a const fn, rather than in concatcp and formatcp.

Defined a `__priv_concatenate` function that uses a `const LEN: usize` parameter,
which does the string concatenation itself, to which concatcp and formatcp delegate to.

Bumped the const_format_proc_macros version to 0.2.14

Updated the github CI config.

Updated root module, and readme to mention this new use of the "const genericss" feature