Added str_get macro, improved index checking in macros.
Improved index checking in macros:
- check that indices are at char boundaries
- use dummy valid indices when the user passes an invalid one,
  allowing the error mssage to be printed from a separate constant.
- define an error type that prints an error message in a sepaate constant to avoid cascading errors

Added a __str_methods::str_indexing module,
with shared types for preparing the passed in range for use, and detecting invalid indices.

Rewrote `str_splice` and `std_index` to use the new way of asserting that the index is valid.
Now instead of having a cascade of error whenever the index is wrong,
there's at most one error per macro invocation.